Update Storybook from 6.2.9 to 6.3.6
Updates the `package.json` and `yarn.lock` files to use the latest version of Storybook.

Update was completed using the `npx sb upgrade` command.

I noticed that it was out of date when building the [Aria Live Regions Addon](https://github.com/chantastic/aria-live-storybook-addon).

I didn't notice recommendations for making this type of contribution. I'm happy to modify it as appropriate.